requirements.txt
paho-mqtt
pvporcupine
vosk
pyaudio
python-dotenv
sounddevice
noisereduce 